Split out local matches and comatches
